Create flipper.py
flipper.py contains 2 methods:

flipbitmask(), which takes a bitmask to be displayed by the hat and flips
it upside-down.

flipped(), which takes a "flipmask" and applies it to a bitmask if it
determines the flipmask is valid for the given bitmask.

Both of these will be used to "flip" the display (aka convert the
bitmasks written to the display to their upside-down counterparts in
real-time).